Allow specifying keyFunction
This allows specifying a custom keyFunction

Use case is I have a listener that unfurls JIRA tickets and I want it to
only do this once per minute max for a given JIRA ticket, but users
should be allowed to use it with different JIRA tickets. Without the
keyFunction, it uses the regex as the key and so different messages
matching the same regex are all treated the same.

I'm using this as follows:

```coffeescript
  regex = /\b[a-zA-Z]{2,12}-[0-9]{1,10}\b/
  listener_options =
    rateLimits:
      minPeriodMs:60000
      keyFunction: (context) -> context.response.message.text

  robot.hear regex, listener_options, (msg) ->
    return if msg.message.user.name.match(new RegExp(ignored_users, "gi"))
```